requirements.txt
torch
pillow
numpy
diffusers
transformers
accelerate
safetensors